Test cases for implicit and explicit string conversions - contributed by Sudharsan V
svn path=/trunk/mcs/; revision=39736

+++ mcs/mbas/Test/tests/ExpConversionSingleToBoolA1.vb
@@ -1,6 +1,6 @@
 Module ExpConversionDoubletoBool
 	Sub Main()
-		Dim a As Long = -4.940656E-12
+		Dim a As Single = -4.940656E-12
 		Dim b as Boolean
 		b = CBool(a)
 		if b <> True
